pyproject.toml
[tool.poetry]
name = "newshub"
version = "0.1.0"
description = "Collaborative news sharing platform"
authors = ["Mathieu Deous <mat.deous@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
django-cors-headers = "^4.3.1"
django-bootstrap5 = "^24.2"
django-mail-panel = "^4.0.4"
gunicorn = "^22.0.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
psycopg2 = "^2.9.9"
whitenoise = "^6.7.0"
django-unfold = "^0.34.0"
celery = "^5.4.0"
redis = "^5.0.7"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.6.0"
feedparser = "^6.0.11"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
black = "^24.4.2"
faker = "^26.0.0"
watchdog = { extras = ["watchmedo"], version = "^4.0.1" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
